修正助教建議

+++ src/views/front/Cart.jsx
@@ -1,6 +1,7 @@
 import axios from "axios";
 import { useEffect, useState } from "react";
 import { currency } from "/src/utils/filter";
+import { Link } from "react-router";
 
 const API_Base = import.meta.env.VITE_API_BASE;
 const API_Path = import.meta.env.VITE_API_PATH;
@@ -30,7 +31,8 @@ function Cart() {
         `${API_Base}/api/${API_Path}/cart/${cartId}`,
         { data },
       );
-      alert(res.data.message);
+
+      // alert(res.data.message);
       // 重新取得購物車資料
       const cartRes = await axios.get(`${API_Base}/api/${API_Path}/cart`);
       setCart(cartRes.data.data);
@@ -40,6 +42,7 @@ function Cart() {
   };
   // 刪除特定 item
   const delCartItem = async (cartId) => {
+    if (!window.confirm("確定要刪除嗎？")) return;
     try {
       const res = await axios.delete(
         `${API_Base}/api/${API_Path}/cart/${cartId}`,
@@ -54,6 +57,7 @@ function Cart() {
   };
   // 刪除全部 item
   const delAllCartItem = async () => {
+    if (!window.confirm("確定要刪除嗎？")) return;
     try {
       const res = await axios.delete(`${API_Base}/api/${API_Path}/carts`);
       alert("成功清空購物車");
@@ -64,91 +68,97 @@ function Cart() {
       console.log(error.response);
     }
   };
-  // 確認是否執行指令
-  const checkoutDel = () => {
-    if (!window.confirm("確定要刪除嗎？")) return;
-    delAllCartItem();
-  };
 
   return (
     <div className="container">
       <h2>購物車列表</h2>
-      <div className="text-end mt-4">
-        <button
-          type="button"
-          className="btn btn-outline-danger"
-          onClick={() => checkoutDel()}
-        >
-          清空購物車
-        </button>
-      </div>
-      <table className="table">
-        <thead>
-          <tr>
-            <th scope="col">品名</th>
-            <th scope="col">數量/單位</th>
-            <th scope="col" className="text-center">
-              小計
-            </th>
-            <th scope="col"></th>
-          </tr>
-        </thead>
-        <tbody>
-          {cart?.carts?.map((cartItem) => (
-            <tr key={cartItem.id}>
-              <th scope="row" className="col-4">
-                {cartItem.product.title}
-              </th>
-              <td className="col-3">
-                <div className="input-group mb-3">
-                  <input
-                    type="number"
-                    className="form-control"
-                    aria-label="Sizing example input"
-                    aria-describedby="inputGroup-sizing-default"
-                    defaultValue={cartItem.qty}
-                    onChange={(e) => {
-                      e.target.value > 1
-                        ? updateQty(
-                            cartItem.id,
-                            cartItem.product_id,
-                            Number(e.target.value),
-                          )
-                        : delCartItem(cartItem.id);
-                    }}
-                  />
-                  <span
-                    className="input-group-text"
-                    id="inputGroup-sizing-default"
-                  >
-                    {cartItem.product.unit}
-                  </span>
-                </div>
-              </td>
-              <td className="col-3 text-center">
-                {currency(cartItem.final_total)}
-              </td>
-              <td className="col-2 text-end">
-                <button
-                  type="button"
-                  className="btn btn-outline-danger btn-sm"
-                  onClick={() => delCartItem(cartItem.id)}
-                >
-                  刪除
-                </button>
-              </td>
-            </tr>
-          ))}
-        </tbody>
-        <tfoot>
-          <tr className="h4">
-            <td className="text-end" colSpan="3">
-              總計
-            </td>
-            <td className="text-end">{currency(cart.final_total)}</td>
-          </tr>
-        </tfoot>
-      </table>
+      {cart?.carts?.length === 0 ? (
+        <div className="my-5">
+          <h3 className="mb-3">目前購物車沒有商品</h3>
+          <button type="button" className="btn btn-outline-primary btn-lg">
+            <Link to="/products">前往選購</Link>
+          </button>
+        </div>
+      ) : (
+        <>
+          <div className="text-end mt-4">
+            <button
+              type="button"
+              className="btn btn-outline-danger"
+              onClick={() => delAllCartItem()}
+            >
+              清空購物車
+            </button>
+          </div>
+          <table className="table">
+            <thead>
+              <tr>
+                <th scope="col">品名</th>
+                <th scope="col">數量/單位</th>
+                <th scope="col" className="text-center">
+                  小計
+                </th>
+                <th scope="col"></th>
+              </tr>
+            </thead>
+            <tbody>
+              {cart?.carts?.map((cartItem) => (
+                <tr key={cartItem.id}>
+                  <th scope="row" className="col-4">
+                    {cartItem.product.title}
+                  </th>
+                  <td className="col-3">
+                    <div className="input-group mb-3">
+                      <input
+                        type="number"
+                        className="form-control"
+                        aria-label="Sizing example input"
+                        aria-describedby="inputGroup-sizing-default"
+                        defaultValue={cartItem.qty}
+                        onChange={(e) => {
+                          e.target.value > 0
+                            ? updateQty(
+                                cartItem.id,
+                                cartItem.product_id,
+                                Number(e.target.value),
+                              )
+                            : delCartItem(cartItem.id);
+                        }}
+                      />
+                      <span
+                        className="input-group-text"
+                        id="inputGroup-sizing-default"
+                      >
+                        {cartItem.product.unit}
+                      </span>
+                    </div>
+                  </td>
+                  <td className="col-3 text-center">
+                    {currency(cartItem.final_total)}
+                  </td>
+                  <td className="col-2 text-end">
+                    <button
+                      type="button"
+                      className="btn btn-outline-danger btn-sm"
+                      onClick={() => delCartItem(cartItem.id)}
+                    >
+                      刪除
+                    </button>
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+            <tfoot>
+              <tr className="h4">
+                <td className="text-end" colSpan="3">
+                  總計
+                </td>
+                <td className="text-end">{currency(cart.final_total)}</td>
+              </tr>
+            </tfoot>
+          </table>
+        </>
+      )}
     </div>
   );
 }
